Significant overhaul of Token class layer
With the goal of supporting separate refresh and access tokens, this
commit makes the following changes:

* The `Token` class can no longer be instantiated directly.  Only
subclasses of `Token` are meant to be instantiated.  Subclasses of
`Token` must define the `token_type` and `lifetime` class attributes to
allow instantiation.

* `Token` subclass instances now set a `current_time` attribute during
instantiation which is meant to help when doing sequences of related
time operations.  The `set_exp` and `check_exp` methods have also been
modified to use this attribute's value wherever they previously used the
value of `datetime.utcnow()`.

* `Token` subclasses now check for the presence of a token type claim
during instantiation which describes the intended use of the token.  If
this claim is missing or its value doesn't match what is expected, an
exception is raised.

* `Token` subclass instances which are not built from an existing,
encoded token now automatically set a default value for the "exp" claim
and also set a value for a new token type claim.  The token type claim
is meant to help Simple JWT determine the intended use of a token.

* `Token.for_user` no longer sets values for the "exp" and "refresh_exp"
claims.  This is now done in the `__init__` method of `Token` classes
and subclasses.

* The functionality which was formerly represented by the `Token` class
is now represented by the `SlidingToken` class which inherits from
`Token`.

* Two other subclasses of `Token` have been added: `AccessToken` and
`RefreshToken`.  These two classes will be used to implement the popular
pattern of having short-lived "access" tokens which are used for
authorization and longer-lived "refresh" tokens which are used to obtain
access tokens and which may also be checked against a token blacklist.